Fixed Unmarshalling of `daysSinceMaintenance`
Previously, `daysSinceMaintenance` was incorrectly interpreted as an integer, which lead to all kinds of craziness. Now, it is correctly unmarshalled as a Double.

+++ MekHQ/src/mekhq/campaign/unit/Unit.java
@@ -1943,7 +1943,7 @@ public class Unit implements ITechnology {
                 } else if (wn2.getNodeName().equalsIgnoreCase("daysActivelyMaintained")) {
                     retVal.daysActivelyMaintained = Integer.parseInt(wn2.getTextContent());
                 } else if (wn2.getNodeName().equalsIgnoreCase("daysSinceMaintenance")) {
-                    retVal.daysSinceMaintenance = Integer.parseInt(wn2.getTextContent());
+                    retVal.daysSinceMaintenance = Double.parseDouble(wn2.getTextContent());
                 } else if (wn2.getNodeName().equalsIgnoreCase("mothballTime")) {
                     retVal.mothballTime = Integer.parseInt(wn2.getTextContent());
                 } else if (wn2.getNodeName().equalsIgnoreCase("astechDaysMaintained")) {
